Share canonical docstring section kinds
